+++ projects/ccf-rui/src/app/modules/registration-modal/registration-content/registration-content.component.scss
