+++ projects/ccf-rui/src/app/modules/registration-modal/registration-modal/registration-modal.component.spec.ts
