+++ projects/ccf-rui/src/app/modules/registration-modal/registration-modal/registration-modal.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { RegistrationModalComponent } from './registration-modal.component';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
-import { RegistrationContentModule } from './registration-content.module';
+import { RegistrationContentModule } from '../registration-content/registration-content.module';
 
 @NgModule({
   declarations: [RegistrationModalComponent],
